moved classes to the separate files and made them exportable
