+++ index.html
@@ -6,6 +6,6 @@
     <title>Document</title>
 </head>
 <body>
-    <script src="lesson_6_hw.js"></script>
+    <script src="lesson_8_hw.js"></script>
 </body>
 </html>
