+++ modules/pm.js
@@ -1,0 +1,9 @@
+import { Employee } from "./employee.js";
+
+export class PM extends Employee { 
+    #jobType = 'PM'; 
+  
+    getJobType() { 
+      return `I am a ${this.#jobType}`;
+    }; 
+  }
