+++ package.json
@@ -10,5 +10,6 @@
   "license": "ISC",
   "devDependencies": {
     "cypress": "^13.3.2"
-  }
+  },
+  "type": "module"
 }
